Refactor views/screens
