+++ client/src/screens/DeleteItem.tsx
