+++ client/src/screens/EmptyContent.tsx
